feat : Errors.swift 파일에 에러 작성
